Update the scoring script
Main update include:
The service now is expecting every element in string format, and one sample input.

The return value would be a JSON string like this:
{"result": [[0.44709373159178445, 0.5529062684082156]]}
and the first value would be the possibility that the patient won't be admitted, while the second value is the possibility that the patient will be admitted. They will add up to 1. We can use a threshold of 0.5 for now (depending how the hospital thinks about it, they can set the threshold differently to be more aggressive or more conservative)